LayoutTests/crypto/workers/subtle/ec-generate-export-spki-key-expected.txt

[Worker] Test exporting an ECDH public key with SPKI format in workers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/ec-generate-export-spki-key.js
[Worker] Generating a key pair...
[Worker] Exporting the public key...
PASS [Worker] publicKey.byteLength is 91
PASS successfullyParsed is true

TEST COMPLETE
